requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Config / Environment ───────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Azure Blob Storage ─────────────────────────────────────────────────────────
azure-storage-blob==12.19.1

# ── ML / Embeddings (PyTorch CPU only — torch installed separately in Dockerfile)
transformers==4.40.2
sentence-transformers==2.7.0
faiss-cpu==1.8.0

# ── LangChain text splitter ────────────────────────────────────────────────────
langchain-text-splitters==0.0.2

# ── Data processing ────────────────────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.2
openpyxl==3.1.2          # .xlsx
xlrd==2.0.1              # legacy .xls
odfpy==1.4.1             # .ods
tqdm==4.66.4

# ── Document parsing ───────────────────────────────────────────────────────────
pypdf==4.2.0             # PDF (primary)
pymupdf==1.24.3          # PDF fallback (PyMuPDF / fitz) — also handles scanned PDFs
python-docx==1.1.2       # .docx
python-pptx==0.6.23      # .pptx
beautifulsoup4==4.12.3   # HTML, XML, ODT, EPUB
lxml==5.2.2              # XML parser for BeautifulSoup
striprtf==0.0.26         # .rtf
pyyaml==6.0.1            # .yaml / .yml
tomli==2.0.1             # .toml (Python < 3.11 fallback)
markdown==3.6            # .md utilities

# ── Google Drive ───────────────────────────────────────────────────────────────
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0

# ── SharePoint (MSAL) ──────────────────────────────────────────────────────────
msal==1.28.0

# ── HTTP / Utilities ───────────────────────────────────────────────────────────
requests==2.32.3
httpx==0.27.0
orjson==3.10.3
